requirements.txt
# Streamlit - Webアプリケーションフレームワーク
streamlit>=1.28.0

# Pillow - 画像処理ライブラリ
Pillow>=10.0.0

# 追加の依存関係（必要に応じて）
# numpy>=1.21.0  # 数値計算（将来的に必要になった場合）